small changes for user stories

+++ library/models.py
@@ -68,11 +68,17 @@ class Book(Document):
 
 
 class ReferenceBook(Book):
-    pass
+    title = models.CharField(max_length=250)
+    library = models.ForeignKey(Library, on_delete=models.DO_NOTHING, related_name='reference')
+    authors = models.ManyToManyField(Author, related_name='reference')
+    keywords = models.ManyToManyField(Keyword, related_name='reference')
 
 
 class AudioVideo(Document):
-    pass
+    title = models.CharField(max_length=250)
+    library = models.ForeignKey(Library, on_delete=models.DO_NOTHING, related_name='audio video')
+    authors = models.ManyToManyField(Author, related_name='audio video')
+    keywords = models.ManyToManyField(Keyword, related_name='audio video')
 
 
 class Editor(models.Model):
@@ -80,8 +86,16 @@ class Editor(models.Model):
     second_name = models.CharField(max_length=250)
 
 
+
 class Journal(models.Model):
     title = models.CharField(max_length=250)
+    library = models.ForeignKey(Library, on_delete=models.DO_NOTHING, related_name='journals')
+    authors = models.ManyToManyField(Author, related_name='journals')
+    price_value = models.IntegerField()
+    keywords = models.ManyToManyField(Keyword, related_name='journals')
+
+    def booking_period(self, user):
+        return datetime.timedelta(weeks=2)
 
 
 class Issue(models.Model):
@@ -261,10 +275,12 @@ class Librarian(User):  # (User,UserCard)
 
     def remove_copy(self, document, count):
         removable = Copy.objects.get(document=document)
-        if removable.number > count:
-            removable.number -= count
-        else:
-            removable.delete()
+        for copy in removable:
+            if copy.is_checked_out == False:
+                if removable.number > count:
+                    removable.number -= count
+                else:
+                    removable.delete()
 
     def patron_information(self, id):
         try:
